Modulação do código e alteração na validação de senha
